flyer/logo updates
Updated flyer v0.001 to reflect the correct times (7, 7:45, & 9).
Made the logo into a vector ai and eps file. Made a high ppi png too.
Collected all of the logo stuff and moved 'em into their own logo
folder. This may have temporarily broken the links in some of the other
files but it is easy enough to update the link.