Contribute page created
put up the html for the contribute page. No styling for the page has
been completed, so it's still pretty ugly. Sorry. Obviously, the form
fields are going to need some integration with your backend database.
Also, If I haven't mentioned it already,: the search filed on all fo
the pages is going to need some serverside code as well. But, you're
awesome so I'm sure you knew that.